integration/requirements.txt
supervisor==4.2.2
gunicorn==20.1.0
flask==2.0.1
python-dateutil==2.8.2
psycopg2==2.9.1
salt-pepper==0.7.6
jira==3.0.1
